Reorganize config directory structure with inline Zod defaults
- Move ConfigManager to src/config/manager.ts
- Create src/config/metadata.ts for CLI/env metadata (CONFIG_METADATA)
- Create src/config/constants.ts for internal-only defaults (retry, tokens)
- Inline all config defaults in Zod schema (single source of truth)
- Delete src/defaults.ts (no longer needed)
- Remove redundant DEFAULT_* fallbacks in server (serverConfig has defaults)
- Fix type assertion in cliGenerator.ts (use ConfigKey instead of any)
- Derive valid categories from CONFIG_METADATA in types.ts
