+++ src/cli/config.ts
@@ -1,3 +1,8 @@
 // Re-export shared config system
-export { config, ConfigManager, type SparkConfig, type SparkConfigResolved } from "../config.js";
+export {
+  config,
+  ConfigManager,
+  type SparkConfig,
+  type SparkConfigResolved,
+} from "../config/index.js";
 export { getConfigDefaults } from "../config/schema.js";
