+++ test/cli/commands/run.test.ts
@@ -1,12 +1,14 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import { Command } from "commander";
 import { createRunCommand } from "../../../src/cli/commands/run.js";
-import {
-  DEFAULT_PROVIDER,
-  DEFAULT_BROWSER,
-  DEFAULT_MAX_ITERATIONS,
-  DEFAULT_MAX_VALIDATION_ATTEMPTS,
-} from "../../../src/defaults.js";
+import { getConfigDefaults } from "../../../src/config/schema.js";
+
+// Get defaults from schema
+const schemaDefaults = getConfigDefaults();
+const DEFAULT_PROVIDER = schemaDefaults.provider;
+const DEFAULT_BROWSER = schemaDefaults.browser;
+const DEFAULT_MAX_ITERATIONS = schemaDefaults.max_iterations;
+const DEFAULT_MAX_VALIDATION_ATTEMPTS = schemaDefaults.max_validation_attempts;
 
 // Mock all the dependencies
 vi.mock("../../../src/webAgent.js", () => ({
@@ -27,12 +29,25 @@ vi.mock("../../../src/browser/playwrightBrowser.js", () => ({
   }),
 }));
 
-vi.mock("../../../src/config.js", () => ({
+// Mock the config manager to avoid fs dependencies
+vi.mock("../../../src/config/manager.js", () => ({
   config: {
     get: vi.fn(),
     getConfig: vi.fn(() => ({})),
   },
+  ConfigManager: vi.fn(),
 }));
+
+vi.mock("../../../src/config/index.js", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("../../../src/config/index.js")>();
+  return {
+    ...actual,
+    config: {
+      get: vi.fn(),
+      getConfig: vi.fn(() => ({})),
+    },
+  };
+});
 
 vi.mock("../../../src/provider.js", () => ({
   createAIProvider: vi.fn(() => ({})),
@@ -57,11 +72,19 @@ vi.mock("../../../src/cli/utils.js", () => ({
   parseResourcesList: vi.fn((resources) => resources.split(",")),
 }));
 
-// Mock fs module
-vi.mock("fs", () => ({
-  mkdirSync: vi.fn(),
-  writeFileSync: vi.fn(),
-}));
+// Mock fs module - include all functions used by manager.ts
+vi.mock("fs", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("fs")>();
+  return {
+    ...actual,
+    default: actual,
+    mkdirSync: vi.fn(),
+    writeFileSync: vi.fn(),
+    // Type assertion needed for overloaded fs functions in test mocks
+    readFileSync: vi.fn().mockReturnValue("{}") as typeof actual.readFileSync,
+    existsSync: vi.fn().mockReturnValue(false) as typeof actual.existsSync,
+  };
+});
 
 // Mock WebAgentEventEmitter
 vi.mock("../../../src/events.js", () => ({
@@ -78,7 +101,7 @@ vi.mock("../../../src/events.js", () => ({
 
 import { WebAgent } from "../../../src/webAgent.js";
 import { PlaywrightBrowser } from "../../../src/browser/playwrightBrowser.js";
-import { config } from "../../../src/config.js";
+import { config } from "../../../src/config/index.js";
 import { createAIProvider } from "../../../src/provider.js";
 import { ChalkConsoleLogger } from "../../../src/loggers/chalkConsole.js";
 import { JSONConsoleLogger } from "../../../src/loggers/json.js";
@@ -106,18 +129,9 @@ describe("CLI Run Command", () => {
     command = createRunCommand();
     vi.clearAllMocks();
 
-    // Set up default config mock
-    mockConfig.get.mockImplementation((key: string, defaultValue?: any) => {
-      const defaults: Record<string, any> = {
-        browser: "firefox",
-        headless: false,
-        block_resources: "media,manifest",
-        max_iterations: 50,
-        max_validation_attempts: 3,
-        bypass_csp: false,
-        logger: "console",
-      };
-      return defaults[key] ?? defaultValue;
+    // Set up default config mock - use schema defaults
+    mockConfig.get.mockImplementation((key: string) => {
+      return schemaDefaults[key as keyof typeof schemaDefaults];
     });
   });
 
@@ -221,11 +235,11 @@ describe("CLI Run Command", () => {
       expect(mockPlaywrightBrowser).toHaveBeenCalledWith(
         expect.objectContaining({
           browser: "firefox",
-          blockAds: true, // Default from --no-block-ads option
-          blockResources: ["media", "manifest"], // Default from config
+          blockAds: true,
+          blockResources: ["media", "manifest"],
           pwEndpoint: undefined,
           headless: false,
-          bypassCSP: false, // Default from config
+          bypassCSP: true, // Schema default
           proxyServer: undefined,
           proxyUsername: undefined,
           proxyPassword: undefined,
@@ -431,11 +445,11 @@ describe("CLI Run Command", () => {
       expect(mockPlaywrightBrowser).toHaveBeenCalledWith(
         expect.objectContaining({
           browser: "firefox",
-          blockAds: true, // Default from --no-block-ads option
-          blockResources: ["media", "manifest"], // Default from config
+          blockAds: true,
+          blockResources: ["media", "manifest"],
           pwEndpoint: undefined,
           headless: false,
-          bypassCSP: false, // Default from config
+          bypassCSP: true, // Schema default
           proxyServer: "http://proxy.company.com:8080",
           proxyUsername: "user",
           proxyPassword: "pass",
@@ -508,10 +522,9 @@ describe("CLI Run Command", () => {
     });
 
     it("should handle invalid browser option", async () => {
-      // Commander validates enum choices and throws InvalidArgumentError
       const args = ["--browser", "invalid", "test task"];
 
-      // Commander's .choices() validation will throw before our custom validation
+      // Commander's .choices() validation throws with "Allowed choices are" message
       await expect(command.parseAsync(args, { from: "user" })).rejects.toThrow(
         /Allowed choices are/i,
       );
